+++ src/app/Employer/dashboard/page.tsx
@@ -1,57 +1,32 @@
-import AccessDenied from "@/components/Employer/AccessDenied";
-import InfoCards from "@/components/Employer/HomeSection/InfoCards";
-import Interviews from "@/components/Employer/HomeSection/Interviews";
-import PostJobs from "@/components/Employer/HomeSection/PostJobs";
-// import { OrganizationSwitcher } from "@clerk/nextjs";
-import { auth, clerkClient } from "@clerk/nextjs/server";
-// import Image from "next/image";
-import React from "react";
+import { Suspense } from "react";
+import { GetEmployerSubscriprionDetails } from "@/components/Employer/EmployerSubscriptionInforamtionCheck";
+import EmployerCheckLoader2 from "@/components/Loaders/EmployerCheckLoader2";
+import { auth } from "@clerk/nextjs/server";
+import InformationDashboardOverview from "@/components/Employer/InformationDashboardOverview";
+import PackagePricingEmployer from "@/components/Employer/PackagePricingEmployer";
 
 export default async function page() {
-  const { orgId,userId } = await auth();
-  const client = await clerkClient();
-  const organization = await client.organizations.getOrganization({
-    organizationId: orgId!,
-  });
-  const { data: organisationMemmbers } =await client.organizations.getOrganizationMembershipList({organizationId: orgId!,});
-  const isOrganisationMember = organisationMemmbers.some((member) => member.publicUserData?.userId === userId,);
-  
-  if(!isOrganisationMember){
-      return(
-        <>
-        <AccessDenied/>
-        </>
-      )
-    }
-
+  const { orgId, userId } = await auth();
+  const { isValidSubscription, isOrganisationMember } =
+    await GetEmployerSubscriprionDetails({
+      orgId: orgId as string,
+      userId: userId as string,
+    });
   return (
     <>
-      <section className="px-4">
-        <div className="c">
-          {/* <p className="c">Hi there, welcome to the</p> */}
-          <h2 className="text-3xl font-semibold mb-10">
-            {organization.name} Dashboard
-          </h2>
-          <InfoCards />
-        </div>
-        <div className="flex items-center gap-2">
-          {/* <div className="flex items-center gap-2">
-            <p className="c">Manage {organization.name} profile </p>
-            <OrganizationSwitcher hidePersonal={true} />
-          </div> */}
-         {/* <Image src={organization.imageUrl} alt="Organization Image" width={80} height={80}  className="bg-cover bg-center bg-no-repeat rounded-lg"/>  */}
-        </div>
-        <div className="flex gap-8">
-          <div className="w-full  md:w-6/12 mx-auto bg-slate-100/50 rounded-md my-10 px-4">
-            <h3 className="text-lg font-semibold my-8">Recently Posted Jobs</h3>
-            <PostJobs />
-          </div>
-          <div className="w-full mx-auto bg-slate-100/60 rounded-md my-10 px-4 md:w-6/12">
-            <h3 className="text-lg font-semibold my-8">Interviews Schedule</h3>
-            <Interviews />
-          </div>
-        </div>
-      </section>
+      <Suspense fallback={<EmployerCheckLoader2 />}>
+        <section className="px-4">
+          {isValidSubscription && isOrganisationMember ? (
+            <>
+              <InformationDashboardOverview />
+            </>
+          ) : (
+            <>
+              <PackagePricingEmployer />
+            </>
+          )}
+        </section>
+      </Suspense>
     </>
   );
 }
