+++ src/app/Employer/page.tsx
@@ -1,6 +1,4 @@
-// import { GetCustomerSubscriptionDetailsByCustomerIDFromPaystack, GetEmployerSubscriptionDetailsFromPaystack } from "@/components/Customer/UserSubscriptionInformation";
-// import {  GetEmployerSubscriptionDetailsFromPaystack } from "@/components/Customer/UserSubscriptionInformation";
-// import { subscriptionDetailsProps } from "@/types/subscriptions";
+
 import { GetEmployerSubscriprionDetails } from "@/components/Employer/EmployerSubscriptionInforamtionCheck";
 import InformationDashboardOverview from "@/components/Employer/InformationDashboardOverview";
 import PackagePricingEmployer from "@/components/Employer/PackagePricingEmployer";
@@ -23,12 +21,12 @@ import { Cog } from "lucide-react";
 import { Suspense } from "react";
 
 export default async function page() {
-  // const client = await clerkClient();
+  
   const { isAuthenticated, orgId, userId } = await auth();
-
-  // if(!isAuthenticated || !userId){redirect("/Employer/sign-in")}
-
-
+  
+  
+  
+  
   if (!isAuthenticated) {
     return (
       <div className="grid place-content-center mt-40 h-96">
@@ -36,20 +34,20 @@ export default async function page() {
           <SignIn
             fallbackRedirectUrl={
               process.env
-                .NEXT_PUBLIC_CLERK_SIGN_IN_FALLBACK_REDIRECT_URL_EMPLOYER_ROUTE
+              .NEXT_PUBLIC_CLERK_SIGN_IN_FALLBACK_REDIRECT_URL_EMPLOYER_ROUTE
             }
             signUpFallbackRedirectUrl={
               process.env
-                .NEXT_PUBLIC_CLERK_SIGN_IN_FALLBACK_REDIRECT_URL_EMPLOYER_ROUTE
+              .NEXT_PUBLIC_CLERK_SIGN_IN_FALLBACK_REDIRECT_URL_EMPLOYER_ROUTE
             }
             routing="hash"
-          />
+            />
         </SignInButton>
       </div>
     );
   }
   
-
+  
   if (!orgId) {
     return (
       <>
@@ -76,44 +74,15 @@ export default async function page() {
       </>
     );
   }
-
-//   //Check if the user is a member of the organization
-//   const { data: organisationMemmbers } =
-//     await client.organizations.getOrganizationMembershipList({
-//       organizationId: orgId!,
-//     });
-//   // console.log("ORGANISATION MEMBERS:", organisationMemmbers);
-
-//   // Find the admin who likely owns the subscription
-// const adminMemmber=organisationMemmbers.find((member)=>member.role=="org:admin")
-// const adminEmail=adminMemmber?.publicUserData?.identifier
-// const adminSubscriptionDetails:subscriptionDetailsProps[]=await GetEmployerSubscriptionDetailsFromPaystack(adminEmail as string)
-
-// //check for valid subscription
-// const isValidSubscription=adminSubscriptionDetails?.filter((subscriptionOne)=>subscriptionOne.amount===300000)
-// .some((subscription)=>{return ["active", "attention", "non-renewing", "completed"].includes(subscription.status.toLowerCase(),)})
-
-// //check for organisation members
-//   const isOrganisationMember = organisationMemmbers.some((member) => member.publicUserData?.userId === userId,);
-
-  //check for valid subscription
-  // const employerSubscriptionDetails: subscriptionDetailsProps[] =
-  //   await GetCustomerSubscriptionDetailsByCustomerIDFromPaystack();
-  // // console.log("EMPLOYER SUBSCRIPTION DETAILS:", employerSubscriptionDetails);
-  // const isValidSubscription = employerSubscriptionDetails
-  //   ?.filter((subscriptionOne) => subscriptionOne.amount === 300000)
-  //   .some((subscription) => {
-  //     return ["active", "attention", "non-renewing", "completed"].includes(
-  //       subscription.status.toLowerCase(),
-  //     );
-  //   });
+  
   const {isValidSubscription,isOrganisationMember}=await GetEmployerSubscriprionDetails({orgId:orgId as string, userId:userId as string})
+
 
   return (
     <>
       <Suspense fallback={<EmployerCheckLoader2 />}>
         <section className="px-4">
-          {/* {isValidSubscription || isOrganisationMember ? ( */}
+          
           {isValidSubscription && isOrganisationMember ? (
             <>
               <InformationDashboardOverview />
