+++ src/components/Employer/CandidatesPage/CandidatesSection.tsx
@@ -1,226 +1,32 @@
-import { Button } from "@/components/ui/button";
-import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogHeader,
-  DialogTitle,
-  DialogTrigger,
-} from "@/components/ui/dialog";
 import { GetCandidatesBasedOnJobDepartment } from "@/db/queries/employerQuries";
-import { correctedParsedHTML } from "@/lib/utils";
+import VirtualizedEmployerCandidateList from "./VirtualizedEmployerCandidateList";
 
-import { Calendar, Mail, MapPin, PhoneIcon, UserCircle } from "lucide-react";
-import { google, office365, outlook, yahoo } from "calendar-link";
-import { Link } from "next-view-transitions";
+export default async function CandidatesSection({
+  jobDepartment,
+}: {
+  jobDepartment: string;
+}) {
+  const candidates = await GetCandidatesBasedOnJobDepartment(jobDepartment);
 
-// import Image from "next/image";
-
-export default async function CandidatesSection({jobDepartment}:{jobDepartment:string }) {
-  // const candidates = await GetCandidatesPool();
-  const candidates= await GetCandidatesBasedOnJobDepartment(jobDepartment)
-
-  const currentYear = new Date().getFullYear();
-  const event = {
-    uuid: "",
-    title: "",
-    description: "",
-    start: `${currentYear}`,
-  };
-  
   return (
     <>
       <div className="max-w-7xl mx-auto px-4">
         {candidates.length === 0 ? (
           <div className="grid place-content-center h-96">
-            <p className="text-center text-lg">No candidates have applied for this role yet</p>
+            <p className="text-center text-lg">
+              No candidates have applied for this role yet
+            </p>
           </div>
-        ) : (<>
-        <p className="text-xl font-semibold">{candidates.length} - Applicants for the role</p>
-          {candidates.map((candidate) => (
-            <div
-              key={candidate.id}
-              className="bg-bts-BrownOne rounded-md px-3 py-6 flex justify-between items-center border my-2 "
-            >
-              <div className="flex flex-1 items-start gap-2">
-                
-                <div className="flex-col">
-                  {/* <p className="flex items-center gap-2 text-xs ">
-                    {candidate.email}
-                  </p> */}
-                  <p className="flex items-center gap-2 text-xs ">
-                    Name:
-                  </p>
-                  <p className="">
-                    {candidate.firstName} <span className="px-1"></span>{" "}
-                    {candidate.lastName}
-                  </p>
-                  
-                </div>
-              </div>
-              {/* <p className="w-full flex-1 text-sm font-semibold"> */}
-              <p className="w-full flex-1 flex-col text-sm font-semibold">
-                
-                <span className="block font-thin text-xs">Profession :</span>
-    
-                {candidate.profession}
-              </p>
-
-              <div className="ml-0 lg:ml-20">
-                <Dialog>
-                  <DialogTrigger asChild>
-                    <Button
-                      variant="outline"
-                      className="hover:scale-105 duration-300 transition ease-in"
-                    >
-                      View candidate details
-                    </Button>
-                  </DialogTrigger>
-                  <DialogContent className="w-full md:max-w-300 max-h-[90dvh] overflow-y-auto">
-                    <DialogHeader>
-                      <DialogTitle className="text-center">
-                        Candidates Profile
-                      </DialogTitle>
-                      <DialogDescription></DialogDescription>
-                    </DialogHeader>
-                    <div className="my-10">
-                      <div className="flex items-center justify-between">
-                        <p className="border rounded-xl px-3 py-1 hidden">
-                          Created on {candidate.createdAt.toLocaleDateString()}
-                        </p>
-                        <p className="border rounded-xl px-3 py-1 hidden">
-                          Deadline on {candidate.createdAt.toLocaleDateString()}
-                        </p>
-                      </div>
-                      <div className="border-t-2 pt-2 flex items-center justify-between gap-2">
-                        <div className="flex items-center gap-2">
-                          <UserCircle size={80} />
-                          <div className="flex-col gap-2">
-                            <p className="text-lg font-medium">
-                              <span className="text-xs hidden">First Name:</span>
-                              {candidate.firstName} {candidate.lastName}
-                            </p>
-                          </div>
-                        </div>
-                        <div className="c">
-                          <p className="text-lg font-medium">
-                            <span className="font-thin text-sm">Profession :</span>
-                            {candidate.profession}
-                          </p>
-                        </div>
-                      </div>
-                      <p className="border-b-2  underline-offset-1 mt-10">
-                        Contact Details:
-                      </p>
-                      <div className="my-10 flex items-center justify-between">
-                        <div className="flex flex-col gap-2 text-sm">
-                          <p className="px-4 py-1 flex items-center gap-1">
-                            {" "}
-                            <Mail size={20} className="" />
-                            {candidate.email}
-                          </p>
-                          <p className="px-4 py-1 flex items-center gap-1">
-                            {" "}
-                            <PhoneIcon size={20} className="" />
-                            {candidate.phone}
-                          </p>
-                          <p className="px-4 py-1 flex items-center gap-1">
-                            {" "}
-                            <MapPin size={20} className="" />
-                            {candidate.country}
-                          </p>
-                        </div>
-                        <Dialog>
-                          <DialogTrigger asChild>
-                            <Button className="bg-lime-200 hover:bg-lime-300 text-black">
-                              <Calendar className="mx-1" /> Schedule calendar event
-                              with {candidate.firstName}
-                            </Button>
-                          </DialogTrigger>
-                          <DialogContent className="max-w-4xl mx-auto">
-                            <DialogHeader>
-                              <DialogTitle></DialogTitle>
-                              <DialogDescription className="text-center">
-                                Schedule your sessions with the candidate using your
-                                prefered calendar below
-                              </DialogDescription>
-                            </DialogHeader>
-                            <div className=" flex flex-wrap items-center justify-center gap-4 my-10">
-                            
-                              <Button variant="outline" asChild>
-                                <Link target="_blank" href={google({...event, guests:[candidate.email],})}>
-                                  Google Calendar
-                                </Link>
-                              </Button>
-                              <Button variant="outline" asChild>
-                                <Link target="_blank" href={outlook({...event, guests:[candidate.email]})}>
-                                  Outlook Calendar
-                                </Link>
-                              </Button>
-                              <Button variant="outline" asChild>
-                                <Link target="_blank" href={office365({...event, guests:[candidate.email]})}>
-                                  Office 365 Calendar
-                                </Link>
-                              </Button>
-                              <Button variant="outline" asChild>
-                                <Link target="_blank" href={yahoo({...event, guests:[candidate.email]})}>
-                                  Yahoo Calendar
-                                </Link>
-                              </Button>
-                              
-                            </div>
-                          </DialogContent>
-                        </Dialog>
-                      </div>
-
-    <p className="border-b-2  underline-offset-1 mt-10">
-                      Resume:
-                      </p>
-                      <div className=" max-w-5xl mx-auto  px-2 py-3 prose prose-sm">
-                        <p className="text-center">
-                          Resume Name:{candidate.resumeName}
-                        </p>
-                        <iframe
-                          
-                          src={`${candidate.resumeLink}#view=fitH`}
-                          title={candidate.resumeName as string}
-                          name={candidate.resumeName as string}
-                          // width={900}
-                          // height={900}
-                          className="w-full h-[70dvh] mx-auto"
-                        />
-                      </div>
-                      <p className="border-b-2  underline-offset-1 mt-10">
-                        Career Experience:
-                      </p>
-                      <div
-                        className="my-10 prose prose-sm md:prose-sm"
-                        dangerouslySetInnerHTML={{
-                          __html: correctedParsedHTML(
-                            candidate.workExperience as string,
-                          ),
-                        }}
-                      ></div>
-                      <p className="border-b-2  underline-offset-1 mt-10">
-                        Certifications:
-                      </p>
-                      <div
-                        className="my-10 prose prose-sm md:prose-sm"
-                        dangerouslySetInnerHTML={{
-                          __html: correctedParsedHTML(
-                            candidate.certifications as string,
-                          ),
-                        }}
-                      ></div>
-                    </div>
-                  </DialogContent>
-                </Dialog>
-              </div>
-
+        ) : (
+          <>
+            <p className="text-xl font-semibold">
+              {candidates.length} - Applicants for the role
+            </p>
+            <div className="my-2 py-2 ">
+              <VirtualizedEmployerCandidateList candidates={candidates} />
             </div>
-          ))}
-  
-        </>)}
+          </>
+        )}
       </div>
     </>
   );
